Rename messageformat2_formatted_value to messageformat2_formatting_context and add docs
