+++ icu4c/source/i18n/sources.txt
@@ -98,7 +98,7 @@ messageformat2.cpp
 messageformat2_checker.cpp
 messageformat2_context.cpp
 messageformat2_data_model.cpp
-messageformat2_formatted_value.cpp
+messageformat2_formatting_context.cpp
 messageformat2_function_registry.cpp
 messageformat2_parser.cpp
 messageformat2_serializer.cpp
