+++ icu4c/source/i18n/unicode/messageformat2_function_registry.h
@@ -13,7 +13,7 @@
 #include "unicode/datefmt.h"
 #include "unicode/format.h"
 #include "unicode/messageformat2_data_model.h"
-#include "unicode/messageformat2_formatted_value.h"
+#include "unicode/messageformat2_formatting_context.h"
 #include "unicode/messageformat2_macros.h"
 #include "unicode/numberformatter.h"
 #include "unicode/unistr.h"
